vault/wallets: test-suite rewired
